+++ spotifywebapipython/models/spotifyconnectdevices.py
@@ -1,0 +1,137 @@
+# external package imports.
+from typing import Iterator
+
+# our package imports.
+from ..sautils import export
+from .device import Device
+from .spotifyconnectdevice import SpotifyConnectDevice
+from spotifywebapipython.zeroconfapi import ZeroconfGetInfo, ZeroconfGetInfoAlias
+
+@export
+class SpotifyConnectDevices():
+    """
+    Spotify Connect Devices collection.
+    """
+
+    def __init__(self) -> None:
+        """
+        Initializes a new instance of the class.
+        """
+        self._Items:list[SpotifyConnectDevice] = []
+
+
+    def __getitem__(self, key) -> SpotifyConnectDevice:
+        return self._Presets[key]
+
+
+    def __iter__(self) -> Iterator:
+        return iter(self._Items)
+
+
+    def __len__(self) -> int:
+        return len(self._Items)
+
+
+    def __repr__(self) -> str:
+        return self.ToString()
+
+
+    def __str__(self) -> str:
+        return self.ToString()
+
+
+    @property
+    def Items(self) -> list[SpotifyConnectDevice]:
+        """ 
+        Array of `SpotifyConnectDevice` objects.
+        """
+        return self._Items
+    
+
+    @property
+    def ItemsCount(self) -> int:
+        """ 
+        Number of objects in the `Items` property array.
+        """
+        if self._Items is not None:
+            return len(self._Items)
+        return 0
+    
+
+    def GetDeviceList(self) -> list[Device]:
+        """
+        Returns a list of `Device` objects that can be used to build a selection list
+        of available devices.  
+        
+        Note that the `Device` object has the following properties populated:
+        Id, Name, IsActive, Type.
+        """
+        result:list[Device] = []
+        
+        # process all discovered devices.
+        scDevice:SpotifyConnectDevice
+        for scDevice in self._Items:
+
+            # map device information details.
+            info:ZeroconfGetInfo = scDevice.DeviceInfo
+
+            # create new mock device.
+            device = Device()
+            device.IsActive = info.HasActiveUser
+            device.Type = info.DeviceType
+                
+            # are aliases being used (RemoteName is null if so)?
+            if info.RemoteName is None:
+                    
+                # if aliases are defined, then use the alias details.
+                infoAlias:ZeroconfGetInfoAlias
+                for infoAlias in info.Aliases:
+                    device.Id = infoAlias.Id
+                    device.Name = infoAlias.Name
+
+            else:
+
+                # if no aliases then use the remote name and id.
+                device.Id = info.DeviceId
+                device.Name = info.RemoteName
+
+            # append device to results.
+            result.append(device)
+
+        # sort items on Name property, ascending order.
+        if len(result) > 0:
+            result.sort(key=lambda x: (x.Name or "").lower(), reverse=False)
+
+        return result
+    
+
+    def ToDictionary(self) -> dict:
+        """
+        Returns a dictionary representation of the class.
+        """
+        result:dict = \
+        {
+            'ItemsCount': self.ItemsCount,
+            'Items': [ item.ToDictionary() for item in self._Items ],
+        }
+
+        return result
+        
+           
+    def ToString(self, includeItems:bool=False) -> str:
+        """
+        Returns a displayable string representation of the class.
+        
+        Args:
+            includeItems (bool):
+                True to include the Items collection of objects; otherwise, False
+                to only return base properties.
+        """
+        msg:str = 'Items Count=%s' % (str(self.ItemsCount))
+        
+        if (includeItems):
+            item:SpotifyConnectDevice
+            for item in self._Items:
+                msg = "%s\n- %s" % (msg, item.ToString(True))
+            
+        return msg